refactor(#5022): extract delivery lease authority
