+++ ARCHITECTURE.md
@@ -1061,7 +1061,9 @@ src/
 │   │   ├── catch_up.rs
 │   │   ├── compact_turn_authority.rs
 │   │   ├── completion_footer_metadata.rs
+│   │   ├── delivery_lease_evidence.rs
 │   │   ├── delivery_lease_key.rs
+│   │   ├── delivery_lease_state.rs
 │   │   ├── destructive_cancel_capture.rs
 │   │   ├── destructive_cancel_gate.rs
 │   │   ├── discord_io.rs
